+++ src/pymordial/core/blueprints/vision_device.py
@@ -1,9 +1,8 @@
+import logging
 from abc import ABC, abstractmethod
-from logging import DEBUG, basicConfig, getLogger
-from typing import TYPE_CHECKING
+from typing import Any
 
-if TYPE_CHECKING:
-    from PIL import Image
+logger = logging.getLogger(__name__)
 
 
 class PymordialVisionDevice(ABC):
@@ -13,13 +12,9 @@ class PymordialVisionDevice(ABC):
     pixel color checking, and text extraction/finding.
     """
 
-    logger = getLogger("PymordialVisionDevice")
-    basicConfig(
-        level=DEBUG, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-    )
-
     @abstractmethod
-    def scale_img_to_screen(self) -> "Image.Image":
+    @abstractmethod
+    def scale_img_to_screen(self) -> Any:
         """Scales the reference image to match the current screen resolution.
 
         Returns: